Extract tables from pdf documents

Using the pdftools package, tables in the pdf document are extracted as a dataframe and returned as a list.

Usage

pages_tables(
  x,
  pages,
  path = FALSE,
  rec = FALSE,
  onecol = FALSE,
  delimiter = "\\s{2,}",
  delimiter_table = "\\s{2,}",
  replacement = "|"
)

Arguments

x

Either the text of the pdf read in with the pdftools package or a path for the location of the pdf file.

pages

Numeric vector of pages from which table will be extracted. it can be a single number or a vector of numbers.

path

TRUE/FALSE; An optional path designation for the location of the pdf to be converted to text. The pdftools package is used for this conversion.

rec

TRUE/FALSE indicating whether the table in the pdf or the page has a rectangular dimension.That is, all rows and all columns are of equal length. This is important because reading a table without proper dimension will produce an error.

onecol

TRUE/FALSE indicating whether the pdf file is one column

delimiter

A delimiter used to detect tables. The default is two consecutive blank white spaces.

delimiter_table

A delimiter used to separate table cells. The default value is two consecutive blank white spaces.

replacement

A delimiter used to separate table cells after the replacement of white space is done.

Value

A list containing data.frame for all extracted table

Examples

file <- system.file("extdata", "onecoldata.pdf", package = "pdftableExtractor")

table2 <- pages_tables(file,pages = 19, path = TRUE)
table2
#> [[1]]
#> [[1]][[1]]
#>                                                                                                                                                  X[[i]]
#> 1                                            2011 ]              THE CRIME REDUCING EFFECT OF EDUCATION                                             481
#> 2                                                          Matching OID to ONS population data, we calculated offending rates (per 1,000 population) by
#> 3                                                          age cohort and year, separately for men and women, using Date of Birth and Gender variables.
#> 4  Downloaded from https://academic.oup.com/ej/article/121/552/463/5079723 by University of Iowa Libraries/Serials Acquisitions user on 04 October 2021
#> 5                                                   Criminal offences have been broadly categorised as property crimes (burglary and theft and handling
#> 6                                                      stolen goods) and violent crimes (violence against the person and robbery), using categorisation
#> 7                                                       in the Offence Class Code variable.15 The overall conviction rate we use is the sum of the two.
#> 8                                                               The data structure for men and women, with means of the total conviction rate per 1,000
#> 9                                                               population, as well as property and violent conviction rates, is presented in Table A1.
#> 10                                                                                                                  A.2. General Household Survey (GHS)
#> 11                                                         Our analysis uses GHS data from 1972 to 1996. The survey took place on a calendar year basis
#> 12                                                           from 1972 to 1987, and then moved to financial year. Using month of survey, we matched the
#> 13                                                                       GHS to the OID for England and Wales by age and year on a calendar year basis.
#> 14                                                                                                             We use two education variables from GHS:
#> 15                                                                   (i) Age left school – variable AGELFTS from 1972 to 1982 and AGELFTSC from 1983 to
#> 16                                                                                         1996. Age left school is set to missing if <13 and above 25.
#> 17                                                            (ii) No educational qualifications – derived from variables measuring highest educational
#> 18                                                                            qualification or whether individuals have any educational qualifications.
#> 19                                                                                                                                             Table A2
#> 20                                                                                                          General Household Survey (GHS) Descriptives
#> 21                                                  GHS          Age            Men, no                 Men,              Women, no              Women,
#> 22                                            Year        range         qualifications       age left school       qualifications       age left school
#> 23                                                 1972        18–26             0.44                  15.85                 0.50                 15.69
#> 24                                                 1973        18–27             0.36                  15.76                 0.46                 15.69
#> 25                                                 1974        18–28             0.36                  15.76                 0.44                 15.73
#> 26                                                 1975        18–29             0.35                  15.76                 0.45                 15.72
#> 27                                                 1976        18–30             0.34                  15.83                 0.43                 15.76
#> 28                                                 1977        18–31             0.32                  16.00                 0.40                 15.87
#> 29                                                 1978        18–32             0.30                  16.03                 0.40                 15.95
#> 30                                                 1979        18–33             0.32                  16.01                 0.37                 15.97
#> 31                                                 1980        18–34             0.29                  16.08                 0.37                 16.01
#> 32                                                 1981        18–35             0.30                  16.07                 0.37                 16.02
#> 33                                                 1982        18–36             0.29                  16.10                 0.33                 16.04
#> 34                                                 1983        18–37             0.28                  16.10                 0.33                 16.02
#> 35                                                 1984        18–38             0.28                  16.06                 0.34                 16.01
#> 36                                                 1985        18–39             0.28                  16.10                 0.32                 16.03
#> 37                                                 1986        18–40             0.26                  16.06                 0.32                 16.02
#> 38                                                 1987        18–40             0.25                  16.12                 0.29                 16.07
#> 39                                                 1988        18–40             0.21                  16.27                 0.23                 16.22
#> 40                                                 1989        18–40             0.20                  16.28                 0.23                 16.24
#> 41                                                 1990        19–40             0.20                  16.28                 0.24                 16.25
#> 42                                                 1991        20–40             0.20                  16.31                 0.21                 16.26
#> 43                                                 1992        21–40             0.17                  16.34                 0.18                 16.31
#> 44                                                 1993        22–40             0.16                  16.41                 0.18                 16.36
#> 45                                                 1994        23–40             0.14                  16.40                 0.18                 16.37
#> 46                                                 1995        24–40             0.16                  16.37                 0.17                 16.37
#> 47                                                 1996        25–40             0.15                  16.38                 0.17                 16.36
#> 
#> 

# extract rectangular data
pages_tables(file, pages = 19, path = TRUE, rec = TRUE)
#> [[1]]
#> [[1]][[1]]
#>                                                                                                                                                 X2011..
#> 1                                                          Matching OID to ONS population data, we calculated offending rates (per 1,000 population) by
#> 2                                                          age cohort and year, separately for men and women, using Date of Birth and Gender variables.
#> 3  Downloaded from https://academic.oup.com/ej/article/121/552/463/5079723 by University of Iowa Libraries/Serials Acquisitions user on 04 October 2021
#> 4                                                   Criminal offences have been broadly categorised as property crimes (burglary and theft and handling
#> 5                                                      stolen goods) and violent crimes (violence against the person and robbery), using categorisation
#> 6                                                       in the Offence Class Code variable.15 The overall conviction rate we use is the sum of the two.
#> 7                                                               The data structure for men and women, with means of the total conviction rate per 1,000
#> 8                                                               population, as well as property and violent conviction rates, is presented in Table A1.
#> 9                                                                                                                   A.2. General Household Survey (GHS)
#> 10                                                         Our analysis uses GHS data from 1972 to 1996. The survey took place on a calendar year basis
#> 11                                                           from 1972 to 1987, and then moved to financial year. Using month of survey, we matched the
#> 12                                                                       GHS to the OID for England and Wales by age and year on a calendar year basis.
#> 13                                                                                                             We use two education variables from GHS:
#> 14                                                                   (i) Age left school – variable AGELFTS from 1972 to 1982 and AGELFTSC from 1983 to
#> 15                                                                                         1996. Age left school is set to missing if <13 and above 25.
#> 16                                                            (ii) No educational qualifications – derived from variables measuring highest educational
#> 17                                                                            qualification or whether individuals have any educational qualifications.
#> 18                                                                                                                                             Table A2
#> 19                                                                                                          General Household Survey (GHS) Descriptives
#> 20                                                                                                                                                  GHS
#> 21                                                                                                                                                 Men,
#> 22                                                                                                                                                 Year
#> 23                                                                                                                                      age left school
#> 24                                                                                                                                                 1972
#> 25                                                                                                                                                15.85
#> 26                                                                                                                                                 1973
#> 27                                                                                                                                                15.76
#> 28                                                                                                                                                 1974
#> 29                                                                                                                                                15.76
#> 30                                                                                                                                                 1975
#> 31                                                                                                                                                15.76
#> 32                                                                                                                                                 1976
#> 33                                                                                                                                                15.83
#> 34                                                                                                                                                 1977
#> 35                                                                                                                                                16.00
#> 36                                                                                                                                                 1978
#> 37                                                                                                                                                16.03
#> 38                                                                                                                                                 1979
#> 39                                                                                                                                                16.01
#> 40                                                                                                                                                 1980
#> 41                                                                                                                                                16.08
#> 42                                                                                                                                                 1981
#> 43                                                                                                                                                16.07
#> 44                                                                                                                                                 1982
#> 45                                                                                                                                                16.10
#> 46                                                                                                                                                 1983
#> 47                                                                                                                                                16.10
#> 48                                                                                                                                                 1984
#> 49                                                                                                                                                16.06
#> 50                                                                                                                                                 1985
#> 51                                                                                                                                                16.10
#> 52                                                                                                                                                 1986
#> 53                                                                                                                                                16.06
#> 54                                                                                                                                                 1987
#> 55                                                                                                                                                16.12
#> 56                                                                                                                                                 1988
#> 57                                                                                                                                                16.27
#> 58                                                                                                                                                 1989
#> 59                                                                                                                                                16.28
#> 60                                                                                                                                                 1990
#> 61                                                                                                                                                16.28
#> 62                                                                                                                                                 1991
#> 63                                                                                                                                                16.31
#> 64                                                                                                                                                 1992
#> 65                                                                                                                                                16.34
#> 66                                                                                                                                                 1993
#> 67                                                                                                                                                16.41
#> 68                                                                                                                                                 1994
#> 69                                                                                                                                                16.40
#> 70                                                                                                                                                 1995
#> 71                                                                                                                                                16.37
#> 72                                                                                                                                                 1996
#> 73                                                                                                                                                16.38
#>    THE.CRIME.REDUCING.EFFECT.OF.EDUCATION            X481
#> 1                                                        
#> 2                                                        
#> 3                                                        
#> 4                                                        
#> 5                                                        
#> 6                                                        
#> 7                                                        
#> 8                                                        
#> 9                                                        
#> 10                                                       
#> 11                                                       
#> 12                                                       
#> 13                                                       
#> 14                                                       
#> 15                                                       
#> 16                                                       
#> 17                                                       
#> 18                                                       
#> 19                                                       
#> 20                                    Age         Men, no
#> 21                              Women, no          Women,
#> 22                                  range  qualifications
#> 23                         qualifications age left school
#> 24                                  18–26            0.44
#> 25                                   0.50           15.69
#> 26                                  18–27            0.36
#> 27                                   0.46           15.69
#> 28                                  18–28            0.36
#> 29                                   0.44           15.73
#> 30                                  18–29            0.35
#> 31                                   0.45           15.72
#> 32                                  18–30            0.34
#> 33                                   0.43           15.76
#> 34                                  18–31            0.32
#> 35                                   0.40           15.87
#> 36                                  18–32            0.30
#> 37                                   0.40           15.95
#> 38                                  18–33            0.32
#> 39                                   0.37           15.97
#> 40                                  18–34            0.29
#> 41                                   0.37           16.01
#> 42                                  18–35            0.30
#> 43                                   0.37           16.02
#> 44                                  18–36            0.29
#> 45                                   0.33           16.04
#> 46                                  18–37            0.28
#> 47                                   0.33           16.02
#> 48                                  18–38            0.28
#> 49                                   0.34           16.01
#> 50                                  18–39            0.28
#> 51                                   0.32           16.03
#> 52                                  18–40            0.26
#> 53                                   0.32           16.02
#> 54                                  18–40            0.25
#> 55                                   0.29           16.07
#> 56                                  18–40            0.21
#> 57                                   0.23           16.22
#> 58                                  18–40            0.20
#> 59                                   0.23           16.24
#> 60                                  19–40            0.20
#> 61                                   0.24           16.25
#> 62                                  20–40            0.20
#> 63                                   0.21           16.26
#> 64                                  21–40            0.17
#> 65                                   0.18           16.31
#> 66                                  22–40            0.16
#> 67                                   0.18           16.36
#> 68                                  23–40            0.14
#> 69                                   0.18           16.37
#> 70                                  24–40            0.16
#> 71                                   0.17           16.37
#> 72                                  25–40            0.15
#> 73                                   0.17           16.36
#> 
#>